[#551] minor cleanup re-positioning of deprecated package
