+++ gradoop-examples/src/main/java/org/gradoop/examples/AbstractRunner.java
@@ -27,13 +27,13 @@ import org.gradoop.flink.io.impl.csv.CSVDataSink;
 import org.gradoop.flink.io.impl.csv.CSVDataSource;
 import org.gradoop.flink.io.impl.csv.indexed.IndexedCSVDataSink;
 import org.gradoop.flink.io.impl.csv.indexed.IndexedCSVDataSource;
-import org.gradoop.flink.io.impl.json.JSONDataSink;
-import org.gradoop.flink.io.impl.json.JSONDataSource;
+import org.gradoop.flink.io.impl.deprecated.json.JSONDataSink;
+import org.gradoop.flink.io.impl.deprecated.json.JSONDataSource;
 import org.gradoop.flink.model.api.epgm.GraphCollection;
 import org.gradoop.flink.model.api.epgm.LogicalGraph;
 import org.gradoop.flink.util.GradoopFlinkConfig;
-import org.gradoop.utils.converter.deprecated.sources.OldCSVDataSource;
-import org.gradoop.utils.converter.deprecated.sources.OldIndexedCSVDataSource;
+import org.gradoop.flink.io.impl.deprecated.logicalgraphcsv.LogicalGraphCSVDataSource;
+import org.gradoop.flink.io.impl.deprecated.logicalgraphcsv.LogicalGraphIndexedCSVDataSource;
 
 import java.io.File;
 import java.io.IOException;
@@ -224,10 +224,10 @@ public abstract class AbstractRunner {
       return new CSVDataSource(directory, config);
     case "indexed":
       return new IndexedCSVDataSource(directory, config);
-    case "oldcsv":
-      return new OldCSVDataSource(directory, config);
-    case "oldindexed":
-      return new OldIndexedCSVDataSource(directory, config);
+    case "lgcsv":
+      return new LogicalGraphCSVDataSource(directory, config);
+    case "lgindexed":
+      return new LogicalGraphIndexedCSVDataSource(directory, config);
     default:
       throw new IllegalArgumentException("Unsupported format: " + format);
     }
