+++ gradoop-examples/src/main/java/org/gradoop/utils/converter/GradoopFormatConverter.java
@@ -28,7 +28,7 @@ public class GradoopFormatConverter extends AbstractRunner implements ProgramDes
    * Converts a graph from one gradoop format to another.
    *
    * args[0] - path to input graph
-   * args[1] - format of input graph (oldcsv, oldindexed, csv, json, indexed)
+   * args[1] - format of input graph (lgcsv, lgindexed, csv, json, indexed)
    * args[2] - path to output graph
    * args[3] - format of output graph (csv, json, indexed)
    *
