+++ gradoop-flink/src/main/java/org/gradoop/flink/io/impl/deprecated/json/JSONBase.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.gradoop.flink.io.impl.json;
+package org.gradoop.flink.io.impl.deprecated.json;
 
 import org.gradoop.flink.util.GradoopFlinkConfig;
