+++ gradoop-flink/src/main/java/org/gradoop/flink/io/impl/deprecated/json/functions/EdgeToJSON.java
@@ -13,44 +13,48 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.gradoop.flink.io.impl.json.functions;
+package org.gradoop.flink.io.impl.deprecated.json.functions;
 
 import org.apache.flink.api.java.io.TextOutputFormat;
 import org.codehaus.jettison.json.JSONException;
 import org.codehaus.jettison.json.JSONObject;
-import org.gradoop.flink.io.impl.json.JSONConstants;
-import org.gradoop.common.model.impl.pojo.Vertex;
+import org.gradoop.flink.io.impl.deprecated.json.JSONConstants;
+import org.gradoop.common.model.impl.pojo.Edge;
 
 /**
- * Converts a vertex into the following format:
+ * Converts an edge into the following format:
  * <p>
  * {
  * "id":0,
- * "data":{"name":"Alice","gender":"female","age":42},
- * "meta":{"label":"Employee","graphs":[0,1,2,3]}
+ * "source":1,
+ * "target":2,
+ * "data":{"since":2015},
+ * "meta":{"label":"friendOf","graphs":[0,1,2,3]}
  * }
  *
- * @param <V> EPGM vertex type
+ * @param <E> EPGM edge type
  */
-public class VertexToJSON<V extends Vertex>
+public class EdgeToJSON<E extends Edge>
   extends EntityToJSON
-  implements TextOutputFormat.TextFormatter<V> {
+  implements TextOutputFormat.TextFormatter<E> {
 
   /**
-   * Creates a JSON string representation from a given vertex.
+   * Creates a JSON string representation from a given edge.
    *
-   * @param v vertex
+   * @param e edge
    * @return JSON string representation
    */
   @Override
-  public String format(V v) {
+  public String format(E e) {
     JSONObject json = new JSONObject();
     try {
-      json.put(JSONConstants.IDENTIFIER, v.getId());
-      json.put(JSONConstants.DATA, writeProperties(v));
-      json.put(JSONConstants.META, writeGraphElementMeta(v));
-    } catch (JSONException e) {
-      e.printStackTrace();
+      json.put(JSONConstants.IDENTIFIER, e.getId());
+      json.put(JSONConstants.EDGE_SOURCE, e.getSourceId());
+      json.put(JSONConstants.EDGE_TARGET, e.getTargetId());
+      json.put(JSONConstants.DATA, writeProperties(e));
+      json.put(JSONConstants.META, writeGraphElementMeta(e));
+    } catch (JSONException ex) {
+      ex.printStackTrace();
     }
     return json.toString();
   }
