+++ gradoop-flink/src/main/java/org/gradoop/flink/io/impl/deprecated/json/functions/JSONToEntity.java
@@ -13,13 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.gradoop.flink.io.impl.json.functions;
+package org.gradoop.flink.io.impl.deprecated.json.functions;
 
 import com.google.common.collect.Maps;
 import org.codehaus.jettison.json.JSONArray;
 import org.codehaus.jettison.json.JSONException;
 import org.codehaus.jettison.json.JSONObject;
-import org.gradoop.flink.io.impl.json.JSONConstants;
+import org.gradoop.flink.io.impl.deprecated.json.JSONConstants;
 import org.gradoop.common.model.impl.id.GradoopId;
 import org.gradoop.common.model.impl.id.GradoopIdSet;
