+++ gradoop-flink/src/main/java/org/gradoop/flink/io/impl/deprecated/json/functions/package-info.java
@@ -16,4 +16,4 @@
 /**
  * Contains functions to transform EPGM entities into JSON and vice versa.
  */
-package org.gradoop.flink.io.impl.json.functions;
+package org.gradoop.flink.io.impl.deprecated.json.functions;
