+++ gradoop-flink/src/main/java/org/gradoop/flink/io/impl/deprecated/json/package-info.java
@@ -16,4 +16,4 @@
 /**
  * Contains all classes related to json input and output to Flink.
  */
-package org.gradoop.flink.io.impl.json;
+package org.gradoop.flink.io.impl.deprecated.json;
