+++ gradoop-flink/src/main/java/org/gradoop/flink/io/impl/deprecated/logicalgraphcsv/CSVLineToEdge.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.gradoop.utils.converter.deprecated.sources;
+package org.gradoop.flink.io.impl.deprecated.logicalgraphcsv;
 
 import org.gradoop.common.model.api.entities.EPGMEdgeFactory;
 import org.gradoop.common.model.impl.id.GradoopId;
@@ -27,7 +27,7 @@ import org.gradoop.flink.io.impl.csv.metadata.MetaData;
  *
  * The string needs to be encoded in the following format:
  *
- * edge-id;source-id;target-id;edge-label;value_1|value_2|...|value_n
+ * {@code edge-id;source-id;target-id;edge-label;value_1|value_2|...|value_n}
  */
 public class CSVLineToEdge extends CSVLineToElement<Edge> {
   /**
