+++ gradoop-flink/src/test/java/org/gradoop/flink/io/impl/deprecated/json/JSONIOTest.java
